Solutions: task1008. Задания 1006, 1007 пока пропустил.

+++ 09 Приведение примитивных типов расширение и сужение/task1008/src/task1008/Solution.java
@@ -21,7 +21,8 @@ public class Solution {
     public static void main(String[] args) {
         int a = 50;
         int b = 17;
-        double d = (float) a / b;
+//        double d =  (float) a / b; // 2.941176414489746 -базовый вариант
+        double d =  (double) a / b; // 2.9411764705882355 -правильный вариант
         System.out.println(d);
     }
 }
